+++ plugins/modules/consul_connect_intention.py
@@ -13,28 +13,28 @@ from ..module_utils.utils import del_none, is_subset
 
 def run_module():
     # define available arguments/parameters a user can pass to the module
-    module_args = dict(
-        state=dict(type="str", choices=["present", "absent"], default="present"),
-        url=dict(type="str", required=True, fallback=(env_fallback, ["CONSUL_HTTP_ADDR"])),
-        validate_certs=dict(type="bool", default=True),
-        connection_timeout=dict(type="int", default=10),
-        management_token=dict(
-            type="str",
-            required=True,
-            no_log=True,
-            fallback=(env_fallback, ["CONSUL_HTTP_TOKEN"]),
-        ),
-        source=dict(type="str", required=True),
-        destination=dict(type="str", required=True),
-        description=dict(type="str"),
-        action=dict(type="str"),
-        permissions=dict(type="list"),
-    )
+    module_args = {
+        "state": {"type": "str", "choices": ["present", "absent"], "default": "present"},
+        "url": {"type": "str", "required": True, "fallback": (env_fallback, ["CONSUL_HTTP_ADDR"])},
+        "validate_certs": {"type": "bool", "default": True},
+        "connection_timeout": {"type": "int", "default": 10},
+        "management_token": {
+            "type": "str",
+            "required": True,
+            "no_log": True,
+            "fallback": (env_fallback, ["CONSUL_HTTP_TOKEN"]),
+        },
+        "source": {"type": "str", "required": True},
+        "destination": {"type": "str", "required": True},
+        "description": {"type": "str"},
+        "action": {"type": "str"},
+        "permissions": {"type": "list"},
+    }
 
     # seed the final result dict in the object. Default nothing changed ;)
-    result = dict(
-        changed=False,
-    )
+    result = {
+        "changed": False,
+    }
 
     # the AnsibleModule object
     module = AnsibleModule(argument_spec=module_args, supports_check_mode=False)
@@ -48,21 +48,20 @@ def run_module():
     )
 
     desired_intention_body = del_none(
-        dict(
-            SourceType="consul",
-            Description=module.params.get("description"),
-            Action=module.params.get("action"),
-            Permissions=module.params.get("permissions"),
-        )
+        {
+            "SourceType": "consul",
+            "Description": module.params.get("description"),
+            "Action": module.params.get("action"),
+            "Permissions": module.params.get("permissions"),
+        }
     )
 
-    if module.params.get("state") == "absent":
-        if existing_intention is not None:
-            consul.delete_connect_intention(
-                source=module.params.get("source"),
-                destination=module.params.get("destination"),
-            )
-            result["changed"] = True
+    if module.params.get("state") == "absent" and existing_intention is not None:
+        consul.delete_connect_intention(
+            source=module.params.get("source"),
+            destination=module.params.get("destination"),
+        )
+        result["changed"] = True
 
     if module.params.get("state") == "present":
         if existing_intention is None or not is_subset(desired_intention_body, existing_intention):
